+++ tests/test_responder.py
@@ -2,8 +2,10 @@ import concurrent
 
 import pytest
 import yaml
+import random
 import responder
 import requests
+import string
 import io
 
 from starlette.responses import PlainTextResponse
@@ -334,9 +336,7 @@ def test_schema_generation():
     from marshmallow import Schema, fields
 
     api = responder.API(
-        title="Web Service",
-        openapi="3.0",
-        allowed_hosts=["testserver", ";"]
+        title="Web Service", openapi="3.0.2", allowed_hosts=["testserver", ";"]
     )
 
     @api.schema("Pet")
@@ -361,18 +361,35 @@ def test_schema_generation():
     dump = yaml.safe_load(r.content)
 
     assert dump
-    assert dump["openapi"] == "3.0"
+    assert dump["openapi"] == "3.0.2"
 
 
 def test_documentation():
     import responder
     from marshmallow import Schema, fields
 
+    description = "This is a sample server for a pet store."
+    terms_of_service = "http://example.com/terms/"
+    contact = {
+        "name": "API Support",
+        "url": "http://www.example.com/support",
+        "email": "support@example.com",
+    }
+    license = {
+        "name": "Apache 2.0",
+        "url": "https://www.apache.org/licenses/LICENSE-2.0.html",
+    }
+
     api = responder.API(
         title="Web Service",
-        openapi="3.0",
+        version="1.0",
+        openapi="3.0.2",
         docs_route="/docs",
-        allowed_hosts=["testserver", ";"]
+        description=description,
+        terms_of_service=terms_of_service,
+        contact=contact,
+        license=license,
+        allowed_hosts=["testserver", ";"],
     )
 
     @api.schema("Pet")
@@ -424,13 +441,23 @@ def test_cookies(api):
     def cookies(req, resp):
         resp.media = {"cookies": req.cookies}
         resp.cookies["sent"] = "true"
+        resp.set_cookie(
+            "hello",
+            "world",
+            expires=123,
+            path="/",
+            max_age=123,
+            secure=False,
+            httponly=True,
+        )
 
     r = api.requests.get(api.url_for(cookies), cookies={"hello": "universe"})
     assert r.json() == {"cookies": {"hello": "universe"}}
     assert "sent" in r.cookies
+    assert "hello" in r.cookies
 
     r = api.requests.get(api.url_for(cookies))
-    assert r.json() == {"cookies": {"sent": "true"}}
+    assert r.json() == {"cookies": {"hello": "world", "sent": "true"}}
 
 
 @pytest.mark.xfail
@@ -441,11 +468,11 @@ def test_sessions(api):
         resp.media = resp.session
 
     r = api.requests.get(api.url_for(view))
-    assert "Responder-Session" in r.cookies
+    assert api.session_cookie in r.cookies
 
     r = api.requests.get(api.url_for(view))
     assert (
-        r.cookies["Responder-Session"]
+        r.cookies[api.session_cookie]
         == '{"hello": "world"}.r3EB04hEEyLYIJaAXCEq3d4YEbs'
     )
     assert r.json() == {"hello": "world"}
@@ -465,28 +492,28 @@ def test_file_uploads(api):
     async def upload(req, resp):
 
         files = await req.media("files")
-        files["hello"] = files["hello"].decode("utf-8")
-        resp.media = {"files": files}
+        result = {}
+        result["hello"] = files["hello"]["content"].decode("utf-8")
+        result["not-a-file"] = files["not-a-file"].decode("utf-8")
+        resp.media = {"files": result}
 
     world = io.StringIO("world")
-    data = {"hello": world}
+    data = {"hello": ("hello.txt", world, "text/plain"), "not-a-file": b"data only"}
     r = api.requests.post(api.url_for(upload), files=data)
-    assert r.json() == {"files": {"hello": "world"}}
+    assert r.json() == {"files": {"hello": "world", "not-a-file": "data only"}}
 
 
 def test_500(api):
-    def catcher(request, exc):
-        return PlainTextResponse("Suppressed error", 500)
-
-    api.app.add_exception_handler(ValueError, catcher)
-
     @api.route("/")
     def view(req, resp):
         raise ValueError
 
-    r = api.requests.get(api.url_for(view))
+    dumb_client = responder.api.TestClient(
+        api, base_url="http://;", raise_server_exceptions=False
+    )
+    r = dumb_client.get(api.url_for(view))
     assert not r.ok
-    assert r.content == b"Suppressed error"
+    assert r.status_code == responder.status_codes.HTTP_500
 
 
 def test_404(api):
@@ -503,8 +530,7 @@ def test_kinda_websockets(api):
         await ws.close()
 
 
-@pytest.mark.xfail
-def test_startup(api, session):
+def test_startup(api):
     who = [None]
 
     @api.route("/{greeting}")
@@ -512,19 +538,12 @@ def test_startup(api, session):
         resp.text = f"{greeting}, {who[0]}!"
 
     @api.on_event("startup")
-    async def asd():
+    async def run_startup():
         who[0] = "world"
-        print("startup")
 
-    @api.on_event("cleanup")
-    async def asd():
-        print("cleanup")
-
-    pool = concurrent.futures.ThreadPoolExecutor(max_workers=2)
-    f = pool.submit(api.run)
-
-    r = requests.get(f"http://localhost:5042/hello")
-    assert r.text == "hello, world!"
+    with api.requests as session:
+        r = session.get(f"http://;/hello")
+        assert r.text == "hello, world!"
 
 
 def test_redirects(api, session):
@@ -553,24 +572,22 @@ def test_session_thoroughly(api, session):
     r = session.get(api.url_for(get))
     assert r.json() == {"session": {"hello": "world"}}
 
-def test_before_response(api, session):
 
+def test_before_response(api, session):
     @api.route("/get")
     def get(req, resp):
         resp.media = req.session
-
 
     @api.route(before_request=True)
     def before_request(req, resp):
         resp.headers["x-pizza"] = "1"
 
     r = session.get(api.url_for(get))
-    assert 'x-pizza' in r.headers
+    assert "x-pizza" in r.headers
+
 
 def test_allowed_hosts():
-    api = responder.API(
-        allowed_hosts=[";", "tenant.;"]
-    )
+    api = responder.API(allowed_hosts=[";", "tenant.;"])
 
     @api.route("/")
     def get(req, resp):
@@ -595,9 +612,7 @@ def test_allowed_hosts():
     r = api.session(base_url="http://unkown_tenant.;").get(api.url_for(get))
     assert r.status_code == 400
 
-    api = responder.API(
-        allowed_hosts=["*.;"]
-    )
+    api = responder.API(allowed_hosts=["*.;"])
 
     @api.route("/")
     def get(req, resp):
@@ -617,3 +632,154 @@ def test_allowed_hosts():
     api._session = None
     r = api.session(base_url="http://tenant2.;").get(api.url_for(get))
     assert r.status_code == 200
+
+
+def create_asset(static_dir, name=None, parent_dir=None):
+    if name is None:
+        name = random.choices(string.ascii_letters, k=6)
+        # :3
+        ext = random.choices(string.ascii_letters, k=2)
+        name = f"{name}.{ext}"
+
+    if parent_dir is None:
+        parent_dir = static_dir
+    else:
+        parent_dir = static_dir.mkdir(parent_dir)
+
+    asset = parent_dir.join(name)
+    asset.write("body { color: blue; }")
+    return asset
+
+
+def test_staticfiles(tmpdir):
+    static_dir = tmpdir.mkdir("static")
+
+    asset1 = create_asset(static_dir)
+    parent_dir = "css"
+    asset2 = create_asset(static_dir, name="asset2", parent_dir=parent_dir)
+
+    api = responder.API(static_dir=str(static_dir))
+    session = api.session()
+
+    static_route = api.static_route
+
+    # ok
+    r = session.get(f"{static_route}/{asset1.basename}")
+    assert r.status_code == api.status_codes.HTTP_200
+
+    r = session.get(f"{static_route}/{parent_dir}/{asset2.basename}")
+    assert r.status_code == api.status_codes.HTTP_200
+
+    # Asset not found
+    r = session.get(f"{static_route}/not_found.css")
+    assert r.status_code == api.status_codes.HTTP_404
+
+    # Not found on dir listing
+    r = session.get(f"{static_route}")
+    assert r.status_code == api.status_codes.HTTP_404
+
+    r = session.get(f"{static_route}/{parent_dir}")
+    assert r.status_code == api.status_codes.HTTP_404
+
+
+def test_staticfiles_custom_route(tmpdir):
+    static_dir = tmpdir.mkdir("static")
+    static_route = "custom/static/route/"
+
+    asset = create_asset(static_dir)
+
+    api = responder.API(static_dir=str(static_dir), static_route=static_route)
+    session = api.session()
+
+    # Check
+    assert api.static_route == "/custom/static/route"
+
+    static_route = api.static_route
+
+    # ok
+    r = session.get(f"{static_route}/{asset.basename}")
+    assert r.status_code == api.status_codes.HTTP_200
+
+    # Asset not found
+    r = session.get(f"{static_route}/not_found.css")
+    assert r.status_code == api.status_codes.HTTP_404
+
+    # Not found on dir listing
+    r = session.get(f"{static_route}")
+    assert r.status_code == api.status_codes.HTTP_404
+
+
+def test_response_html_property(api):
+    @api.route("/")
+    def view(req, resp):
+        resp.html = "<h1>Hello !</h1>"
+
+        assert resp.content == "<h1>Hello !</h1>"
+        assert resp.mimetype == "text/html"
+
+    r = api.requests.get(api.url_for(view))
+    assert r.content == b"<h1>Hello !</h1>"
+    assert r.headers["Content-Type"] == "text/html"
+
+
+def test_response_text_property(api):
+    @api.route("/")
+    def view(req, resp):
+        resp.text = "<h1>Hello !</h1>"
+
+        assert resp.content == "<h1>Hello !</h1>"
+        assert resp.mimetype == "text/plain"
+
+    r = api.requests.get(api.url_for(view))
+    assert r.content == b"<h1>Hello !</h1>"
+    assert r.headers["Content-Type"] == "text/plain"
+
+
+def test_stream(api, session):
+    async def shout_stream(who):
+        for c in who.upper():
+            yield c
+
+    @api.route("/{who}")
+    async def greeting(req, resp, *, who):
+
+        resp.stream(shout_stream, who)
+
+    r = session.get("/morocco")
+    assert r.text == "MOROCCO"
+
+    @api.route("/")
+    async def home(req, resp):
+        # Raise when it's not an async generator
+        with pytest.raises(AssertionError):
+
+            def foo():
+                pass
+
+            res.stream(foo)
+
+        with pytest.raises(AssertionError):
+
+            async def foo():
+                pass
+
+            res.stream(foo)
+
+        with pytest.raises(AssertionError):
+
+            def foo():
+                yield "oopsie"
+
+            res.stream(foo)
+
+
+def test_empty_req_text(api):
+    content = "It's working"
+
+    @api.route("/")
+    async def home(req, resp):
+        await req.text
+        resp.text = content
+
+    r = api.requests.post("/")
+    assert r.text == content